feat: ios: redesign create derived key action
