+++ ios/PolkadotVault/Modals/Errors/ErrorBottomModalViewModel.swift
@@ -150,27 +150,6 @@ struct ErrorBottomModalViewModel {
         )
     }
 
-    static func derivedKeysInfo(_ action: @escaping @autoclosure () -> Void = {}())
-        -> ErrorBottomModalViewModel {
-        ErrorBottomModalViewModel(
-            title: Localizable.CreateDerivedKey.InfoModal.DerivedKeys.title.string,
-            content: Localizable.CreateDerivedKey.InfoModal.DerivedKeys.content.string,
-            secondaryAction: .init(label: Localizable.CreateDerivedKey.InfoModal.DerivedKeys.action.key, action: action)
-        )
-    }
-
-    static func derivationPathsInfo(_ action: @escaping @autoclosure () -> Void = {}())
-        -> ErrorBottomModalViewModel {
-        ErrorBottomModalViewModel(
-            title: Localizable.CreateDerivedKey.InfoModal.DerivationPaths.title.string,
-            content: Localizable.CreateDerivedKey.InfoModal.DerivationPaths.content.string,
-            secondaryAction: .init(
-                label: Localizable.CreateDerivedKey.InfoModal.DerivationPaths.action.key,
-                action: action
-            )
-        )
-    }
-
     static func noNetworksAvailable(_ action: @escaping @autoclosure () -> Void = {}()) -> ErrorBottomModalViewModel {
         ErrorBottomModalViewModel(
             title: Localizable.KeyDetails.Error.NoNetworks.title.string,
