+++ ios/PolkadotVault/Screens/DerivedKey/Subviews/DerivationPathNameView.swift
@@ -117,15 +117,23 @@ struct DerivationPathNameView: View {
             viewModel.onAppear()
             focusedPath = true
         }
-        .onReceive(viewModel.dismissViewRequest) { _ in
-            presentationMode.wrappedValue.dismiss()
-        }
         .fullScreenCover(
             isPresented: $viewModel.isPresentingInfoModal
         ) {
             DerivationMethodsInfoView(
                 viewModel: .init(
                     isPresented: $viewModel.isPresentingInfoModal
+                )
+            )
+            .clearModalBackground()
+        }
+        .fullScreenCover(
+            isPresented: $viewModel.isPresentingConfirmation
+        ) {
+            CreateDerivedKeyConfirmationView(
+                viewModel: .init(
+                    derivationPath: viewModel.unwrappedDerivationPath(),
+                    onCompletion: viewModel.onConfirmationCompletion
                 )
             )
             .clearModalBackground()
@@ -177,8 +185,11 @@ extension DerivationPathNameView {
     final class ViewModel: ObservableObject {
         private weak var navigation: NavigationCoordinator!
         private let createKeyService: CreateDerivedKeyService
-
         private let seedName: String
+        private let onComplete: () -> Void
+        private var skipValidation = false
+        private let cancelBag = CancelBag()
+
         @Published var inputText: String = ""
         @Published var maskedText: String = ""
         @Published var passwordConfirmation: String = ""
@@ -186,10 +197,9 @@ extension DerivationPathNameView {
         @Published var isPasswordValid: Bool = true
         @Published var isMainActionDisabled: Bool = true
         @Published var derivationPathError: String?
-        @Binding var derivationPath: String?
+        @Published var isPresentingConfirmation: Bool = false
+        @Published var derivationPath: String?
         @Binding var networkSelection: NetworkSelection
-        private var skipValidation = false
-        private let cancelBag = CancelBag()
 
         var isEntrySuggestionActive: Bool {
             DerivationPathComponent.allCases.contains { inputText == $0.description } && derivationPathError == nil
@@ -197,24 +207,18 @@ extension DerivationPathNameView {
 
         // State presentatation
         @Published var isPresentingInfoModal: Bool = false
-        @Published var presentableInfoModal: ErrorBottomModalViewModel = .derivationPathsInfo()
-
-        var dismissViewRequest: AnyPublisher<Void, Never> {
-            dismissRequest.eraseToAnyPublisher()
-        }
-
-        private let dismissRequest = PassthroughSubject<Void, Never>()
+        @Published var presentableInfoModal: ErrorBottomModalViewModel!
 
         init(
             seedName: String,
-            derivationPath: Binding<String?>,
             networkSelection: Binding<NetworkSelection>,
-            createKeyService: CreateDerivedKeyService = CreateDerivedKeyService()
+            createKeyService: CreateDerivedKeyService = CreateDerivedKeyService(),
+            onComplete: @escaping () -> Void
         ) {
             self.seedName = seedName
-            _derivationPath = derivationPath
             _networkSelection = networkSelection
             self.createKeyService = createKeyService
+            self.onComplete = onComplete
             subscribeToChanges()
         }
 
@@ -234,7 +238,21 @@ extension DerivationPathNameView {
 
         func onRightNavigationButtonTap() {
             derivationPath = inputText
-            dismissRequest.send()
+            let completion: (Result<Void, Error>) -> Void = { result in
+                switch result {
+                case .success:
+                    self.isPresentingConfirmation = true
+                case let .failure(error):
+                    self.presentableInfoModal = .alertError(message: error.localizedDescription)
+                    self.isPresentingInfoModal = true
+                }
+            }
+            switch networkSelection {
+            case let .network(network):
+                createKeyService.createDerivedKey(seedName, unwrappedDerivationPath(), network.key, completion)
+            case .allowedOnAnyNetwork:
+                createKeyService.createDerivedKeyOnAllNetworks(seedName, unwrappedDerivationPath(), completion)
+            }
         }
 
         func onInfoBoxTap() {
@@ -251,6 +269,11 @@ extension DerivationPathNameView {
 
         func onPasswordedPathTap() {
             inputText.append(DerivationPathComponent.passworded.description)
+        }
+
+        func onConfirmationCompletion() {
+            isPresentingConfirmation = false
+            onComplete()
         }
 
         func validateDerivationPath() {
@@ -332,6 +355,10 @@ extension DerivationPathNameView {
         private func isInitialEntry() -> Bool {
             inputText == DerivationPathComponent.hard.description
         }
+
+        func unwrappedDerivationPath() -> String {
+            derivationPath ?? ""
+        }
     }
 }
 
@@ -341,8 +368,8 @@ extension DerivationPathNameView {
             DerivationPathNameView(
                 viewModel: .init(
                     seedName: "Keys",
-                    derivationPath: .constant("path"),
-                    networkSelection: .constant(.allowedOnAnyNetwork([]))
+                    networkSelection: .constant(.allowedOnAnyNetwork([])),
+                    onComplete: {}
                 )
             )
             .environmentObject(NavigationCoordinator())
